Feat: repository services (Pt. 1) (#892)
Adds the go2proto package into our protobuf generation script, and
automatically generates the protobuf message types. Any struct with the
// @go2proto comment above them in types will be exported as protobuf
messages.

This is part 1, which lays out the various structs and sets up the grpc
services for the worker repo and container repo. The follow up will
implement the actual repository services needed for the worker with the
generated types.

+++ pkg/types/scheduler.go
@@ -67,6 +67,7 @@ func (e *ContainerAlreadyScheduledError) Error() string {
 	return e.Msg
 }
 
+// @go2proto
 type ContainerState struct {
 	ContainerId string          `redis:"container_id" json:"container_id"`
 	StubId      string          `redis:"stub_id" json:"stub_id"`
@@ -79,6 +80,7 @@ type ContainerState struct {
 	Memory      int64           `redis:"memory" json:"memory"`
 }
 
+// @go2proto
 type BuildOptions struct {
 	SourceImage      *string  `json:"source_image"`
 	Dockerfile       *string  `json:"dockerfile"`
@@ -87,6 +89,7 @@ type BuildOptions struct {
 	BuildSecrets     []string `json:"build_secrets"`
 }
 
+// @go2proto
 type ContainerRequest struct {
 	ContainerId       string          `json:"container_id"`
 	EntryPoint        []string        `json:"entry_point"`
@@ -125,9 +128,9 @@ const (
 	ContainerDurationEmissionInterval      time.Duration = 5 * time.Second
 	ContainerResourceUsageEmissionInterval time.Duration = 3 * time.Second
 )
-const ContainerStateTtlSWhilePending int = 600
-const ContainerStateTtlS int = 120
-const WorkspaceQuotaTtlS int = 600
+const ContainerStateTtlSWhilePending int64 = 600
+const ContainerStateTtlS int64 = 120
+const WorkspaceQuotaTtlS int64 = 600
 
 type ErrContainerStateNotFound struct {
 	ContainerId string
@@ -204,6 +207,7 @@ const (
 	CheckpointStatusNotFound         CheckpointStatus = "not_found"
 )
 
+// @go2proto
 type CheckpointState struct {
 	StubId      string           `redis:"stub_id" json:"stub_id"`
 	ContainerId string           `redis:"container_id" json:"container_id"`
